ai-lead-discovery-agent/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
beautifulsoup4==4.12.3
playwright==1.41.2
google-generativeai==0.4.0
openpyxl==3.1.2
python-multipart==0.0.7
aiofiles==23.2.1
lxml==5.1.0
